% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{checkB}
\alias{checkB}
\title{Helper function to confirm that the argument B to several functions is as
expected}
\usage{
checkB(B)
}
\arguments{
\item{B}{Integer or numeric; the number of subsamples. Note: For
sampling.type=="MB" the total number of subsamples will be \code{B}; for
sampling_type="SS" the number of subsamples will be \code{2*B}.}
}
\description{
Helper function to confirm that the argument B to several functions is as
expected
}
\author{
Gregory Faletto, Jacob Bien
}