% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cssr.Rmd.
\name{checkClusters}
\alias{checkClusters}
\title{Helper function to confirm that the argument clusters to several functions is
as expected}
\usage{
checkClusters(clusters, p)
}
\arguments{
\item{clusters}{A named list where each entry is an integer vector of indices
of features that are in a common cluster, as in the output of css or
formatClusters. (The length of list clusters is equal to the number of
clusters.) All identified clusters must be non-overlapping, and all features
must appear in exactly one cluster (any unclustered features should be in
their own "cluster" of size 1).}

\item{p}{The number of features; must be at least as large as the number of
clusters.}
}
\description{
Helper function to confirm that the argument clusters to several functions is
as expected
}
\author{
Gregory Faletto, Jacob Bien
}