% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{checkCssLoopOutput}
\alias{checkCssLoopOutput}
\title{Helper function to confirm that the outputs of the provided feature selection
method are as required.}
\usage{
checkCssLoopOutput(selected, p, feats_on_subsamp)
}
\arguments{
\item{selected}{An integer vector; the indices of the features selected by
the lasso.}

\item{p}{The total number of observed features; all selected features must be
in 1:p.}

\item{feats_on_subsamp}{Integer; the indices of the features considered by
the feature selection method. All selected features must be among these
features.}
}
\description{
Helper function to confirm that the outputs of the provided feature selection
method are as required.
}
\author{
Gregory Faletto, Jacob Bien
}