% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cssr.Rmd.
\name{checkCutoff}
\alias{checkCutoff}
\title{Helper function to confirm that the argument cutoff to several functions is
as expected}
\usage{
checkCutoff(cutoff)
}
\arguments{
\item{cutoff}{Numeric; only those clusters with selection proportions equal
to at least cutoff will be selected by cluster stability selection. Must be
between 0 and 1.}
}
\description{
Helper function to confirm that the argument cutoff to several functions is
as expected
}
\author{
Gregory Faletto, Jacob Bien
}