% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cssr.Rmd.
\name{checkGetClusterPropsOutput}
\alias{checkGetClusterPropsOutput}
\title{Helper function to check the output of getClusterProps}
\usage{
checkGetClusterPropsOutput(
  feat_sel_props,
  p,
  clus_sel_props_p,
  res_n_clusters,
  clusters,
  sampling_type,
  B
)
}
\arguments{
\item{feat_sel_props}{A numeric
vector of length p. The jth entry is the proportion of subsamples in which
feature j was selected by fitfun.}

\item{p}{Integer; the number of features in the X matrix provided to css.}

\item{clus_sel_props_p}{A numeric vector
of length p. The jth entry is the proportion of subsamples in which at least
one member of the cluster containing feature j was selected.}

\item{res_n_clusters}{A binary integer matrix with the same number of rows
as res and length(clusters) columns. res_n_clusters\link{i, j} = 1 if at least
one member of cluster j was selected on subsample i, and 0 otherwise.}

\item{clusters}{A named list where each entry is an integer vector of indices
of features that are in a common cluster, as in the output of formatClusters.
(The length of list clusters is equal to the number of clusters.) All
identified clusters must be non-overlapping, and all features must appear in
exactly one cluster (any unclustered features should be in their own
"cluster" of size 1).}

\item{B}{Integer; half of the subsamples for sampling_type "SS".}
}
\description{
Helper function to check the output of getClusterProps
}
\author{
Gregory Faletto, Jacob Bien
}