% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cssr.Rmd.
\name{checkGetClusterSelMatrixInput}
\alias{checkGetClusterSelMatrixInput}
\title{Helper function to check inputs to getClusterSelMatrix function}
\usage{
checkGetClusterSelMatrixInput(clusters, res)
}
\arguments{
\item{clusters}{A named list where each entry is an integer vector of indices
of features that are in a common cluster, as in the output of formatClusters.
(The length of list clusters is equal to the number of clusters.) All
identified clusters must be non-overlapping, and all features must appear in
exactly one cluster (any unclustered features should be in their own
"cluster" of size 1).}

\item{res}{A binary integer matrix. res\link{i, j} = 1 if feature j was selected
on subsample i and equals 0 otherwise, as in the output of getSelMatrix.
(That is, each row is a selected set.)}
}
\value{
The parameter B, corresponding to half of the subsamples for
sampling_type "SS".
}
\description{
Helper function to check inputs to getClusterSelMatrix function
}
\author{
Gregory Faletto, Jacob Bien
}