% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{checkGetCssPredsInputs}
\alias{checkGetCssPredsInputs}
\title{Helper function to confirm that inputs to the function getCssPreds are as
expected, and modify inputs if needed.}
\usage{
checkGetCssPredsInputs(
  css_results,
  testX,
  weighting,
  cutoff,
  min_num_clusts,
  max_num_clusts,
  trainX,
  trainY
)
}
\arguments{
\item{css_results}{An object of class "cssr" (the output of the function
css).}

\item{testX}{A numeric matrix (preferably) or a data.frame (which will
be coerced internally to a matrix by the function model.matrix) containing
the data that will be used to generate predictions. Must contain the same
features (in the same number of columns) as the matrix provided to css.}

\item{weighting}{Character; determines how to calculate the weights to
combine features from the selected clusters into weighted averages, called
cluster representatives. Must be one of "sparse", "weighted_avg", or
"simple_avg'. For "sparse", all the weight is put on the most frequently
selected individual cluster member (or divided equally among all the clusters
that are tied for the top selection proportion if there is a tie). For
"weighted_avg", the weight used for each cluster member is calculated in
proportion to the individual selection proportions of each feature. For
"simple_avg", each cluster member gets equal weight regardless of the
individual feature selection proportions (that is, the cluster representative
is just a simple average of all the cluster members). See Faletto and Bien
(2022) for details. Default is "weighted_avg".}

\item{cutoff}{Numeric; getCssPreds will make use only of those clusters with
selection proportions equal to at least cutoff. Must be between 0 and 1.
Default is 0 (in which case either all clusters are used, or max_num_clusts
are used, if max_num_clusts is specified).}

\item{min_num_clusts}{Integer or numeric; the minimum number of clusters to
use regardless of cutoff. (That is, if the chosen cutoff returns fewer than
min_num_clusts clusters, the cutoff will be increased until at least
min_num_clusts clusters are selected.) Default is 1.}

\item{max_num_clusts}{Integer or numeric; the maximum number of clusters to
use regardless of cutoff. (That is, if the chosen cutoff returns more than
max_num_clusts clusters, the cutoff will be decreased until at most
max_num_clusts clusters are selected.) Default is NA (in which case
max_num_clusts is ignored).}

\item{trainX}{A numeric matrix (preferably) or a data.frame (which will
be coerced internally to a matrix by the function model.matrix) containing
the data that will be used to estimate the linear model from the selected
clusters. trainX is only necessary to provide if no train_inds were
designated in the css function call to set aside observations for model
estimation (though even if train_inds was provided, trainX and trianY will be
used for model estimation if they are both provided to getCssPreds). Must
contain the same features (in the same number of columns) as the matrix
provided to css, and if the columns of trainX are labeled, the names must
match the variable names provided to css. Default is NA (in which case
getCssPreds uses the observations from the train_inds that were provided to
css to estimate a linear model).}

\item{trainY}{The response corresponding to trainX. Must be a real-valued
response (unlike in the general css setup) because predictions will be
generated by an ordinary least squares model. Must have the same length as
the number of rows of trainX. Like trainX, only needs to be provided if no
observations were set aside for model estimation by the parameter train_inds
in the css function call. Default is NA (in which case getCssPreds uses the
observations from the train_inds that were provided to css).}
}
\value{
A named list with the following elements: \item{trainXProvided}{
Logical; indicates whether a valid trainX input was provided.} \item{trainX}{
The provided trainX matrix, coerced from a data.frame to a matrix if the
provided trainX was a data.frame. (If a valid trainX was not provided, this
output simply passes whatever was provided as trainX.)} \item{testX}{The
provided testX matrix, coerced from a data.frame to a matrix if the provided
testX was a data.frame.} \item{feat_names}{A character vector containing the
column names of testX (if the provided testX had column names). If the
provided testX did not have column names, feat_names will be NA.}
\item{max_num_clusts}{The provided max_num_clusts, coerced to an integer if
needed, and coerced to be less than or equal to the total number of clusters
from the output of css_results.}
}
\description{
Helper function to confirm that inputs to the function getCssPreds are as
expected, and modify inputs if needed.
}
\author{
Gregory Faletto, Jacob Bien
}