% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cssr.Rmd.
\name{checkGetSelectedClustersOutput}
\alias{checkGetSelectedClustersOutput}
\title{Helper function to check that output of getSelectedClusters is as expected}
\usage{
checkGetSelectedClustersOutput(selected_clusts, selected_feats, n_clusters, p)
}
\arguments{
\item{selected_clusts}{A named numeric vector containing the selection
proportions for the selected clusters. The name of each entry is the name of
the corresponding cluster.}

\item{selected_feats}{A named integer vector; the indices of the features
with nonzero weights from all of the selected clusters.}

\item{n_clusters}{Integer; the number of clusters in the data (upper bound
for the length of selected_clusts)}

\item{p}{Integer; number of features in the data (all selected_feats should
be in 1:p)}
}
\description{
Helper function to check that output of getSelectedClusters is as expected
}
\author{
Gregory Faletto, Jacob Bien
}