% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{checkMinNumClusts}
\alias{checkMinNumClusts}
\title{Helper function to confirm that the argument min_num_clusts to several
functions is as expected}
\usage{
checkMinNumClusts(min_num_clusts, p, n_clusters)
}
\arguments{
\item{min_num_clusts}{Integer or numeric; the minimum number of clusters to
use regardless of cutoff. (That is, if the chosen cutoff returns fewer than
min_num_clusts clusters, the cutoff will be increased until at least
min_num_clusts clusters are selected.)}

\item{p}{The number of features; since this is an upper bound on the number
of clusters of features, it is also an upper bound on min_num_clusts.}

\item{n_clusters}{The number of clusters; note that this is an upper bound
on min_num_clusts}
}
\description{
Helper function to confirm that the argument min_num_clusts to several
functions is as expected
}
\author{
Gregory Faletto, Jacob Bien
}