% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cssr.Rmd.
\name{checkNewXProvided}
\alias{checkNewXProvided}
\title{Helper function to confirm that the new X matrix provided to getCssDesign or
getCssPreds matches the characteristics of the X that was provided to css.}
\usage{
checkNewXProvided(trainX, css_results)
}
\arguments{
\item{trainX}{A numeric matrix (preferably) or a data.frame (which will
be coerced internally to a matrix by the function model.matrix). Must contain
the same features (in the same number of columns) as the X matrix provided to
css, and if the columns of trainX are labeled, the names must match the
variable names provided to css. trainX may be omitted if train_inds were
provided to css to set aside observations.}

\item{css_results}{An object of class "cssr" (the output of the function
css).}
}
\value{
A named list with the following elements: \item{newX}{The provided
trainX matrix, coerced from a data.frame to a matrix if the provided trainX
was a data.frame.} \item{newXProvided}{Logical; indicates whether a valid
trainX input was provided.}
}
\description{
Helper function to confirm that the new X matrix provided to getCssDesign or
getCssPreds matches the characteristics of the X that was provided to css.
}
\author{
Gregory Faletto, Jacob Bien
}