% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{checkSamplingType}
\alias{checkSamplingType}
\title{Helper function to confirm that the argument sampling_type to several
functions is as expected}
\usage{
checkSamplingType(sampling_type)
}
\arguments{
\item{sampling_type}{A character vector; either "SS" or "MB". "MB" is not
supported yet.}
}
\description{
Helper function to confirm that the argument sampling_type to several
functions is as expected
}
\author{
Gregory Faletto, Jacob Bien
}