% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{checkSelectedClusters}
\alias{checkSelectedClusters}
\title{Helper function to check operations within getSelectedClusters function}
\usage{
checkSelectedClusters(
  n_sel_clusts,
  min_num_clusts,
  max_num_clusts,
  max_sel_prop
)
}
\arguments{
\item{n_sel_clusts}{The number of selected clusters; should be constrained
by min_num_clusts and max_num_clusts (though it may not be possible to
satisfy both constraints simulteneously, in which case a warning will be
thrown).}

\item{min_num_clusts}{Integer or numeric; the minimum number of clusters to
use regardless of cutoff. (That is, if the chosen cutoff returns fewer than
min_num_clusts clusters, the cutoff will be increased until at least
min_num_clusts clusters are selected.)}

\item{max_num_clusts}{Integer or numeric; the maximum number of clusters to
use regardless of cutoff. (That is, if the chosen cutoff returns more than
max_num_clusts clusters, the cutoff will be decreased until at most
max_num_clusts clusters are selected.) If NA, max_num_clusts is ignored.}

\item{max_sel_prop}{Numeric; the maximum selection proportion observed for
any cluster.}
}
\description{
Helper function to check operations within getSelectedClusters function
}
\author{
Gregory Faletto, Jacob Bien
}