% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{checkWeighting}
\alias{checkWeighting}
\title{Helper function to confirm that the argument weighting to several
functions is as expected}
\usage{
checkWeighting(weighting)
}
\arguments{
\item{weighting}{Character; determines how to calculate the weights to
combine features from the selected clusters into weighted averages, called
cluster representatives. Must be one of "sparse", "weighted_avg", or
"simple_avg'.}
}
\description{
Helper function to confirm that the argument weighting to several
functions is as expected
}
\author{
Gregory Faletto, Jacob Bien
}