% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{checkXInputResults}
\alias{checkXInputResults}
\title{Helper function to confirm that inputs to several functions are as expected,
and modify inputs if needed}
\usage{
checkXInputResults(newx, css_X)
}
\arguments{
\item{newx}{A numeric matrix (preferably) or a data.frame (which will
be coerced internally to a matrix by the function model.matrix) containing
the data that will be used to generate the design matrix of cluster
representatives. Must contain the same features (in the same
number of columns) as the X matrix provided to css, and if the columns of
newX are labeled, the names must match the variable names provided to css.}

\item{css_X}{The X matrix provided to css, as in the output of the css
function (after having been coerced from a data.frame to a matrix by css if
needed).}
}
\value{
A named list with the following elements. \item{feat_names}{A
character vector containing the column names of newx (if the provided newx
had column names). If the provided newx did not have column names, feat_names
will be NA.} \item{newx}{The provided newx matrix, coerced from a data.frame
to a matrix if the provided newx was a data.frame.}
}
\description{
Helper function to confirm that inputs to several functions are as expected,
and modify inputs if needed
}
\author{
Gregory Faletto, Jacob Bien
}