% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{checkY}
\alias{checkY}
\title{Helper function to confirm that the argument y to several functions is
as expected}
\usage{
checkY(y, n)
}
\arguments{
\item{y}{Numeric response vector.}

\item{n}{Number of observations of covariates; should match length of y.}
}
\description{
Helper function to confirm that the argument y to several functions is
as expected
}
\author{
Gregory Faletto, Jacob Bien
}