% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{corFunction}
\alias{corFunction}
\title{Absolute value of sample correlation between two vectors}
\usage{
corFunction(t, y)
}
\arguments{
\item{t}{A numeric or integer vector.}

\item{y}{A numeric or integer vector; must have the same length as t.}
}
\value{
A length-one numeric vector; the absolute value of the sample
correlation between t and y (between 0 and 1), or 0 if either t or y has only
one unique value.
}
\description{
Calculates the absolute value of correlation of t and y. If either input has
only one unique value, returns 0 by definition.
}
\author{
Gregory Faletto, Jacob Bien
}