% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssr-package.R
\docType{package}
\name{cssr-package}
\alias{cssr}
\alias{cssr-package}
\title{Cluster Stability Selection}
\description{
Cluster stability selection is a feature selection method designed to allow
stability selection to work effectively in the presence of highly correlated
features.  It was proposed in \url{https://arxiv.org/abs/2201.00494}.  To learn
more about this package, please visit its website
\url{https://gregfaletto.github.io/cssr-project/}.
}
\seealso{
\code{\link{css}} \code{\link{cssSelect}}
}
\author{
\strong{Maintainer}: Gregory Faletto \email{gregory.faletto@marshall.usc.edu} [copyright holder]

Authors:
\itemize{
  \item Gregory Faletto \email{gregory.faletto@marshall.usc.edu} [copyright holder]
  \item Jacob Bien \email{jbien@usc.edu} [copyright holder]
}

}
\keyword{internal}
