% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cssr.Rmd.
\name{getClusterProps}
\alias{getClusterProps}
\title{Find cluster proportions}
\usage{
getClusterProps(clusters, res, sampling_type)
}
\arguments{
\item{clusters}{A named list where each entry is an integer vector of indices
of features that are in a common cluster, as in the output of formatClusters.
(The length of list clusters is equal to the number of clusters.) All
identified clusters must be non-overlapping, and all features must appear in
exactly one cluster (any unclustered features should be in their own
"cluster" of size 1).}

\item{res}{A binary integer matrix of dimension \code{B} x \code{p} (if sampling_type
== "MB" was provided to ) or \code{2*B} x \code{p} (if sampling_type == "SS").
res\link{i, j} = 1 if feature j was selected on subsample i and equals 0
otherwise, as in the output of getSelMatrix. (That is, each row is a selected
set.)}

\item{sampling_type}{A character vector (either "SS" or "MB"); the sampling
type used for stability selection.}
}
\value{
A named list with four elements. \item{feat_sel_props}{A numeric
vector of length p. The jth entry is the proportion of subsamples in which
feature j was selected by fitfun.} \item{clus_sel_props_p}{A numeric vector
of length p. The jth entry is the proportion of subsamples in which at least
one member of the cluster containing feature j was selected.}
\item{res_clus_p}{An integer matrix of the same dimensions as res.
res_clus_p\link{i, j} = 1 if at least one member of the cluster containing feature
j was selected by fitfun on subsample i, and 0 otherwise.}
\item{res_n_clusters}{A binary integer matrix with the same number of rows
as res and length(clusters) columns. res_n_clusters\link{i, j} = 1 if at least
one member of cluster j was selected on subsample i, and 0 otherwise.}
}
\description{
Given a matrix of feature selection indicator variables and a list of
clusters of features, returns a matrix of cluster indicator variables, as
well as the selection proportions for both features and clusters.
}
\author{
Gregory Faletto, Jacob Bien
}