% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{getNoiseVar}
\alias{getNoiseVar}
\title{Get variance of noise to add to Z in order to yield proxies X with desired
correlations with Z}
\usage{
getNoiseVar(rho)
}
\arguments{
\item{rho}{A numeric vector of desired correlations for each proxy to have
with Z. Note: correlations must be in (0, 1] (positive and at most 1).}
}
\value{
A vector of variances of independent Gaussian random variables to add
to Z in order to yield proxies with the desired correlations with Z.
}
\description{
Get variance of noise to add to Z in order to yield proxies X with desired
correlations with Z
}
\examples{
# Noise variance needed so a proxy Z + N(0, v) attains each target
# correlation with Z (correlations must be in (0, 1] (positive and at most 1)).
getNoiseVar(c(0.9, 0.5, 1))
}
\author{
Gregory Faletto, Jacob Bien
}