% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{protolasso}
\alias{protolasso}
\title{Select features via the protolasso (Reid and Tibshirani 2016)}
\usage{
protolasso(X, y, clusters, nlambda = 100)
}
\arguments{
\item{X}{An n x p numeric matrix (preferably) or a data.frame (which will
be coerced internally to a matrix by the function model.matrix) containing
p >= 2 features/predictors}

\item{y}{The response; A length n numeric (or integer) real-valued vector.}

\item{clusters}{A list of integer vectors; each vector should contain the
indices of a cluster of features (a subset of 1:p). (If there is only one
cluster, clusters can either be a list of length 1 or an integer vector.)
All of the provided clusters must be non-overlapping. Every feature not
appearing in any cluster will be assumed to be unclustered (that is, they
will be treated as if they are in a "cluster" containing only themselves).
Default is list() (so no clusters are specified).}

\item{nlambda}{Integer; the number of lambda values to use in the lasso fit
for the protolasso. Default is 100 (following the default for glmnet). For
now, nlambda must be at least 2 (using a single lambda is not supported).}
}
\value{
A list with three elements. \item{selected_sets}{A list of integer
vectors. Entry k of this list contains a selected set (an integer vector) of
size k yielded by the protolasso (If no set of size k was selected, entry k
will be empty.)} \item{selected_clusts_list}{A list; each element of the list
is a named list of selected clusters. (That is, if a selected set of size k
was yielded by the protolasso, then selected_clusts_list[\link{k}] is a named
list of length k, where each member of the list is an integer vector
of cluster members. In particular, selected_clusts_lists[\link{k}][\link{j}] will be
the cluster that contains feature selected_sets[\link{k}]\link{j}.)} \item{beta}{The
beta output from glmnet when the lasso was estimated on a matrix of
prototypes. (See documentation for the function glmnet from the glmnet
package for details.)}
}
\description{
Select features via the protolasso (Reid and Tibshirani 2016)
}
\references{
Reid, S., & Tibshirani, R. (2016). Sparse regression and marginal
testing using cluster prototypes. \emph{Biostatistics}, 17(2), 364–376.
\url{https://doi.org/10.1093/biostatistics/kxv049}.
}
\author{
Gregory Faletto, Jacob Bien
}