% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runCSSR}
\alias{runCSSR}
\title{Run CSSR Algorithm}
\usage{
runCSSR(alphabet, data, maxLength, isChi, sigLevel, outputPrefix)
}
\arguments{
\item{alphabet}{String representing alphabet}

\item{data}{String representing discrete-state time series data}

\item{maxLength}{Maximum history length}

\item{isChi}{Boolean to indicate whether to perform chi-square test}

\item{sigLevel}{Significant level}

\item{outputPrefix}{Prefix for output files}
}
\description{
Write some docs!
}

