% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cssr.Rmd.
\name{createSubsamples}
\alias{createSubsamples}
\title{Creates lists of subsamples for stability selection.}
\usage{
createSubsamples(n, p, B, sampling_type, prop_feats_remove = 0)
}
\arguments{
\item{n}{Integer or numeric; sample size of the data set.}

\item{p}{Integer or numeric; number of features.}

\item{B}{Integer or numeric; the number of subsamples. Note: For
sampling.type=="MB" the number of lasso fits will be \code{B}; for
sampling_type="SS" the number of lasso fits will be \code{2*B}.}

\item{sampling_type}{A character vector (either "SS" or "MB"); the sampling
type used for stability selection.}

\item{num_feats_remove}{Integer; number of features select automatically on
every iteration. Determined earlier from input prop_feats_remove to css.}
}
\value{
A list of length \code{B} (or \code{2*B} for sampling_type = "SS"). If
prop_feats_remove = 0, each list element is an integer vector of length
floor(\code{n/2}) containing the indices of a subsample of 1:\code{n}. (For
sampling_type=="SS", the last \code{B} subsamples will be complementary pairs of
the first \code{B} subsamples; see Faletto and Bien 2022 or Shah and Samworth 2013
for details.) If prop_feats_remove > 0, each element is a named list with
members "subsample" (same as above) and "feats_to_keep", a logical vector
of length p; feats_to_keep\link{j} = TRUE if feature j is chosen for this
subsample, and false otherwise.
}
\description{
Creates lists of subsamples for stability selection.
}
\references{
Faletto, G., & Bien, J. (2022). Cluster Stability Selection.
\emph{arXiv preprint arXiv:2201.00494}.
\url{https://arxiv.org/abs/2201.00494}. \cr Shah, R. D., & Samworth, R. J.
(2013). Variable selection with error control: Another look at stability
selection. \emph{Journal of the Royal Statistical Society. Series B:
Statistical Methodology}, 75(1), 55–80.
\url{https://doi.org/10.1109/RITA.2014.2302071}.
}
\author{
Gregory Faletto, Jacob Bien
}