% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{cssr-package}
\alias{cssr}
\alias{cssr-package}
\title{Cluster Stability Selection}
\description{
DESCRIBE PACKAGE HERE AND PERHAPS POINT USER TO PKGDOWN WEBSITE ONCE
IT EXISTS.
}
\seealso{
\code{\link{css}}
}
\author{
\strong{Maintainer}: Gregory Faletto \email{gregory.faletto@marshall.usc.edu}

Authors:
\itemize{
  \item Jacob Bien \email{jbien@usc.edu}
}

}
\keyword{internal}