% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cssr.Rmd.
\name{getModelSize}
\alias{getModelSize}
\title{Automated estimation of model size}
\usage{
getModelSize(X, y, clusters)
}
\arguments{
\item{X}{An n x p numeric matrix (preferably) or a data.frame (which will
be coerced internally to a matrix by the function model.matrix) containing
the p >= 2 features/predictors.}

\item{y}{A length-n numeric vector containing the responses; \code{y[i]} is the
response corresponding to observation \code{X[i, ]}. (Note that for the css
function, y does not have to be a numeric response, but for this function,
the underlying selection procedure is the lasso, so y must be a real-valued
response.)}

\item{clusters}{A named list where each entry is an integer vector of indices
of features that are in a common cluster, as in the output of css.
(The length of list clusters is equal to the number of clusters.) All
identified clusters must be non-overlapping, and all features must appear in
exactly one cluster (any unclustered features should be in their own
"cluster" of size 1).}
}
\value{
An integer; the estimated size of the model. The minimum returned
value is 1, even if the lasso with cross-validation chose no features.
}
\description{
This function is used internally by the functions cssSelect and cssPredict
if no selection proportion threshold (cutoff) or maximum number of clusters
(max_num_clusts) is provided as an input by the user and auto_select_size is
TRUE. getModelSize uses the lasso with cross-validation to estimate the
model size, which is then provided as max_num_clusts to cssSelect or
cssPredict. Before using the lasso, in each cluster all features will be
dropped from X except the one feature with the highest marginal correlation
with y, as in the protolasso (Reid and Tibshirani 2016).
}
\references{
Reid, S., & Tibshirani, R. (2016). Sparse regression and marginal
testing using cluster prototypes. \emph{Biostatistics}, 17(2), 364–376.
\url{https://doi.org/10.1093/biostatistics/kxv049}.
}
\author{
Gregory Faletto, Jacob Bien
}