% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{getNoiseVar}
\alias{getNoiseVar}
\title{Get variance of noise to add to Z in order to yield proxies X with desired
correlations}
\usage{
getNoiseVar(cor)
}
\arguments{
\item{cor}{A numeric vector of desired correlations for each proxy to have
with Z. Note: correlations must be positive.}
}
\value{
A vector of variances of independent Gaussian random variables to add
to Z in order to yield proxies with the desired correlations with Z.
}
\description{
Get variance of noise to add to Z in order to yield proxies X with desired
correlations
}
\author{
Gregory Faletto, Jacob Bien
}