% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{getSelMatrix}
\alias{getSelMatrix}
\title{Generates matrix of selection indicators from stability selection.}
\usage{
getSelMatrix(
  x,
  y,
  lambda,
  B,
  sampling_type,
  subsamps_object,
  num_cores,
  fitfun = cssLasso
)
}
\arguments{
\item{x}{an n x p numeric matrix or a data.frame containing the predictors.}

\item{y}{A response vector; can be any response that takes the form of a
length n vector and is used (or not used) by fitfun. Typically (and for
default fitfun = cssLasso), y should be an n-dimensional numeric vector
containing the response.}

\item{lambda}{A tuning parameter or set of tuning parameters that may be used
by the feature selection method. For example, in the default case when
fitfun = cssLasso, lambda is a numeric: the penalty to use for each lasso
fit.}

\item{B}{Integer or numeric; the number of subsamples. Note: For
sampling.type=="MB" the number of lasso fits will be \code{B}; for
sampling_type="SS" the number of lasso fits will be \code{2*B}.}

\item{sampling_type}{A character vector (either "SS" or "MB"); the sampling
type used for stability selection.}

\item{subsamps_object}{A list of length \code{B} (or \code{2*B} if sampling_type="SS"),
where each element is one of the following: \item{subsample}{An integer
vector of size \code{n/2} containing the indices of the observations in the
subsample.} \item{drop_var_input}{A named list containing two elements: one
named "subsample", matching the previous description, and a logical vector
named "feats_to_keep" containing the indices of the features to be
automatically selected.} (The first object is the output of the function
createSubsamples when the provided prop_feats_remove is 0, the default, and
the second object is the output of createSubsamples when prop_feats_remove >
0.)}

\item{num_cores}{Optional; an integer. If using parallel processing, the
number of cores to use for parallel processing (num_cores will be supplied
internally as the mc.cores argument of parallel::mclapply).}

\item{fitfun}{A function that takes in arguments X, y, and lambda and returns
a vector of indices of the columns of X (selected features). Default is
cssLasso, an implementation of the lasso using glmnet.}
}
\value{
A binary integer matrix of dimension \code{B} x \code{p} (if sampling_type ==
"MB") or \code{2*B} x \code{p} (if sampling_type == "SS"). res\link{i, j} = 1 if feature j
was selected on subsample i and equals 0 otherwise. (That is, each row is a
selected set.)
}
\description{
Generates matrix of selection indicators from stability selection.
}
\author{
Gregory Faletto, Jacob Bien
}