% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runCSSR}
\alias{runCSSR}
\title{Run CSSR Algorithm}
\usage{
runCSSR(alphabet, data, maxLength, chiSquare, sigLevel = 0.001)
}
\arguments{
\item{alphabet}{something}

\item{data}{something}

\item{maxLength}{something}

\item{chiSquare}{something}

\item{sigLevel}{something}
}
\description{
Write some docs!
}
