% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runCSSR}
\alias{runCSSR}
\title{Run CSSR Algorithm}
\usage{
runCSSR(alphabet, data, maxLength, isChi, sigLevel, outputPrefix)
}
\arguments{
\item{alphabet}{String representing alphabet}

\item{data}{String representing discrete-state time series data}

\item{maxLength}{Maximum history length}

\item{isChi}{Boolean to indicate whether to perform chi-square test}

\item{sigLevel}{Significant level}

\item{outputPrefix}{Prefix for output files}
}
\value{
a named list
}
\description{
Run CSSR algorithm to creates separate causal states for each history of data
	with a singular probability distribution.  History length
	increases incrementally until cutoff point is reached.  Then
            removes transient states, determinizes remaining states, and
            calculates various metrics for the resulting state machine.
}
